backend/requirements.txt
flask>=3.0
flask-cors>=4.0
torch>=2.2
torchvision>=0.17
Pillow>=10.0
numpy>=1.26
gradio>=4.0